+++ python/h3ronpy/polars/__init__.py
@@ -30,6 +30,7 @@ def _wrap(func, ret_type=None):
 
 
 change_resolution = _wrap(_arrow.change_resolution, ret_type=pl.Series)
+change_resolution_list = _wrap(_arrow.change_resolution, ret_type=pl.Series)
 change_resolution_paired = _wrap(_arrow.change_resolution_paired, ret_type=pl.DataFrame)
 cells_resolution = _wrap(_arrow.cells_resolution, ret_type=pl.Series)
 cells_parse = _wrap(_arrow.cells_parse, ret_type=pl.Series)
@@ -66,6 +67,12 @@ class H3Expr:
 
     def cells_resolution(self) -> pl.Expr:
         return self._expr.map(lambda s: cells_resolution(s)).alias("resolution")
+
+    def change_resolution(self) -> pl.Expr:
+        return self._expr.map(lambda s: change_resolution(s))
+
+    def change_resolution_list(self) -> pl.Expr:
+        return self._expr.map(lambda s: change_resolution_list(s))
 
     def cells_parse(self, set_failing_to_invalid: bool = False) -> pl.Expr:
         return self._expr.map(lambda s: cells_parse(s, set_failing_to_invalid=set_failing_to_invalid)).alias("cell")
@@ -132,6 +139,12 @@ class H3SeriesShortcuts:
     def cells_resolution(self) -> pl.Series:
         return cells_resolution(self._s)
 
+    def change_resolution(self) -> pl.Series:
+        return change_resolution(self._s)
+
+    def change_resolution_list(self) -> pl.Series:
+        return change_resolution_list(self._s)
+
     def cells_parse(self, set_failing_to_invalid: bool = False) -> pl.Series:
         return cells_parse(self._s, set_failing_to_invalid=set_failing_to_invalid)
 
@@ -180,6 +193,7 @@ class H3SeriesShortcuts:
 
 __all__ = [
     change_resolution.__name__,
+    change_resolution_list.__name__,
     change_resolution_paired.__name__,
     cells_resolution.__name__,
     cells_parse.__name__,