+++ src/op/mod.rs
@@ -9,6 +9,7 @@ mod valid;
 
 pub fn init_op_submodule(m: &PyModule) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(resolution::change_resolution, m)?)?;
+    m.add_function(wrap_pyfunction!(resolution::change_resolution_list, m)?)?;
     m.add_function(wrap_pyfunction!(resolution::change_resolution_paired, m)?)?;
     m.add_function(wrap_pyfunction!(resolution::cells_resolution, m)?)?;
     m.add_function(wrap_pyfunction!(neighbor::grid_disk, m)?)?;